+++ source/extensions/omni.isaac.core_nodes/nodes/OgnIsaacComputeOdometry.cpp
@@ -68,7 +68,7 @@ public:
             }
             else
             {
-                db.logError("no prim path found");
+                db.logWarning("no prim path found");
                 return false;
             }