+++ source/extensions/omni.isaac.debug_draw/nodes/OgnIsaacXPrimRadiusVisualizer.cpp
@@ -58,7 +58,7 @@ public:
         }
         else
         {
-            db.logError("no prim path found");
+            db.logWarning("no prim path found");
             return false;
         }
         auto& state = db.internalState<OgnIsaacXPrimRadiusVisualizer>();