+++ source/extensions/omni.isaac.ros2_bridge/nodes/OgnROS2PublishJointState.cpp
@@ -55,7 +55,7 @@ public:
         }
         else
         {
-            db.logError("no prim path found");
+            db.logWarning("no prim path found");
             return false;
         }