+++ source/extensions/omni.isaac.ros2_bridge/nodes/OgnROS2PublishTransformTree.cpp
@@ -83,7 +83,7 @@ public:
             }
             else
             {
-                db.logError("Please specify atleast one target prim for the ROS pose tree component");
+                db.logWarning("Please specify atleast one target prim for the ROS pose tree component");
                 return false;
             }